Version 0.9-10: Fixed/updated parameters of addWorkFile() and updated roxygen documentation
The original purpose of this commit was to add the 'storage' argument to addWorkFile().  While making this change, it was observed that the 'group' and 'token' parameters were not being passed into hiveURL() when constructing the API call, so these were added as well.

This commit also includes roxygen edits to cover undocumented S4 slots, as well as other edits to documentation or code for consistency and clarity (notably, that the storage system may not always be referred to as a "hive", now that it has been rebranded as Operend Core).