Version 0.9-20
This commit fixes an issue resulting from an update to the definition of the Group object, implemented starting in version 0.9-7.  Originally a Group contained only the 'name' field, and listGroups() returned only a character vector of group names for convenience; however, the Group object now also contains a list of Users who belong to that group, and so listGroups() returns either (by default) a data frame or a list of hiveGroup objects.  The edits to addUser() and addWorkFile() in this commit reflect that change and bring them in line with a similar edit to hiveUpdate() in package version 0.9-12.